app/training/requirements.txt
# --- Training & Math (The essentials) ---
pandas
scikit-learn
joblib
pyarrow           # Required to read the .parquet file!
mlflow 
azureml-mlflow    # Required to talk to Azure's MLflow server
azureml-core
